Add files via upload
The related code is the gold deposit prospectivity prediction method based on ensemble learning and multi-model of deep learning network. First, we put 12 original data maps drawn by Arcgis into BWG folder and modify the size of the pixels represented by the split grid in the Data_split.py file and run, it will divide the original image data according to the same size. The split file will be saved in the split folder, the same kind of image data will be in the same name sub-folder. After running Data_process.py file, the program automatically read different sub-folders with the same partition image data, the image after grayscale processing in accordance with a fixed order stacked together and packaged into a_b.npy packets (representing the file with a row and b column area) that are stored in the Datas folder. According to the collected data for the sample division, modify the pixel size parameter of the clipping window and the sliding step parameter of the window in Data_augmentation.py. After running the program, it will read the positive and negative sample data in turn, the a_b.npy data files are split layer by layer and the data is enhanced and repackaged, and the resulting files are stored in the New Data subfolder of the Datas folder. After running the move.py, 80% of the positive and negative samples in the New Data were randomly divided into training set and 20% into test set. Model training. Once the dataset is ready, run annotation.py to generate the bwg_train.txt and bwg_test.txt required for the training. This text record is the training and test samples of the storage path and the corresponding classification label. Then, we run the train.py file to start the model training, and the training log is automatically saved in the logs folder. At last, we run evaluation.py under the eval folder, and the generated evaluation information is stored in the eval_results folder.